cmake support done

+++ include/Scanner/Token.h
@@ -3,7 +3,7 @@
 #include <string>
 #include <iostream>
 #include <any>
-#include "TokenType.h"
+#include "Scanner\TokenType.h"
 
 class Token
 {
@@ -11,7 +11,7 @@ class Token
         TokenType type;
         std::string lexeme;
         std::any literal;
-        int line;
+        size_t line;
 
         Token(TokenType type, const std::string& lexeme, const std::any literal, int line)
         : type(type), lexeme(lexeme), literal(literal), line(line) {}
